Handler: Fix sorting for items and projects
The item_order property for items are only unique per project.
Because of this we need to adjust the sort algorithm to first
sort by project_id, and then by item_order to get the correct
presentatoin sorting.

We were also previously sorting by project_id for projects, but
that doesn't work correctly if you haven't created your projects
in order and you have re-nested them. Sorting by purely item_order
seems to be the right thing to do for projects.